rewardpoints-2.0

+++ src/main/java/com/assignment/charter1/model/Transaction.java
@@ -9,6 +9,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -22,5 +23,5 @@ public class Transaction {
     private Long id;
     private Long customerId;
     private LocalDate transactionDate;
-    private Double transactionAmount;
+    private BigDecimal transactionAmount;
 }
